requirements.txt
aiogram==3.22.0
aiohttp
python-dotenv
